requirements.txt
pyyaml>=6.0
pandas>=1.5.0
rich>=13.0.0
pytest>=7.0.0
